chore: split CosmosWallet to hooks, types
