+++ src/components/Wallet/CosmotWallet.types.ts
@@ -1,0 +1,75 @@
+declare global {
+  interface Window {
+    keplr?: KeplrWallet;
+    crossmark?: CrossmarkWallet;
+  }
+}
+
+interface KeplrChain {
+  chainId: string;
+  chainName: string;
+  rpc: string;
+  rest: string;
+  bip44: {
+    coinType: number;
+  };
+  bech32Config: {
+    bech32PrefixAccAddr: string;
+    bech32PrefixAccPub: string;
+    bech32PrefixValAddr: string;
+    bech32PrefixValPub: string;
+    bech32PrefixConsAddr: string;
+    bech32PrefixConsPub: string;
+  };
+  currencies: Array<{
+    coinDenom: string;
+    coinMinimalDenom: string;
+    coinDecimals: number;
+  }>;
+  feeCurrencies: Array<{
+    coinDenom: string;
+    coinMinimalDenom: string;
+    coinDecimals: number;
+  }>;
+  stakeCurrency: {
+    coinDenom: string;
+    coinMinimalDenom: string;
+    coinDecimals: number;
+  };
+  features: string[];
+}
+
+export interface KeplrSigner {
+  getAccounts(): Promise<Array<{ address: string }>>;
+  signAmino(
+    signerAddress: string,
+    signDoc: Record<string, unknown>
+  ): Promise<Record<string, unknown>>;
+  signDirect(
+    signerAddress: string,
+    signDoc: Record<string, unknown>
+  ): Promise<Record<string, unknown>>;
+}
+
+interface KeplrWallet {
+  enable(chainId: string): Promise<void>;
+  experimentalSuggestChain(chain: KeplrChain): Promise<void>;
+  getOfflineSignerAuto(chainId: string): Promise<KeplrSigner>;
+}
+
+interface CrossmarkWallet {
+  isConnected: boolean;
+  connect(): Promise<void>;
+  disconnect(): Promise<void>;
+}
+
+export interface CosmosWalletState {
+  chainId: string | null;
+  address: string | null;
+  provider: KeplrWallet | null;
+  signer: KeplrSigner | null;
+  setChainId: (chainId: string | null) => void;
+  setAddress: (address: string | null) => void;
+  setProvider: (provider: KeplrWallet | null) => void;
+  setSigner: (signer: KeplrSigner | null) => void;
+}
